dummy for RemarkableCloudAPI
